+++ NixService/DTOs/PurchaseDto.cs
@@ -6,7 +6,7 @@ namespace NixService.DTOs
 {
     public class PurchaseDto
     {
-        public int PaymentMethodNumber { get; set; }
+        public long PaymentMethodNumber { get; set; }
         public decimal Value { get; set; }
         public string Description { get; set; }
